# Conflicts:
#	faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/quran/downloadedSur/DownloadedSurViewModel.kt

+++ admin_panel/src/commonMain/kotlin/net/thechance/mena/admin_panel/data/remote/dto/PagedResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PagedResponse<T>(
     @SerialName("totalElements")
-    val totalElements: Long?,
+    val totalElements: Int?,
     @SerialName("page")
     val page: Int?,
     @SerialName("pageSize")
